vertigo/functions/email-processor/requirements.txt
google-cloud-secret-manager
google-auth
google-auth-oauthlib
google-api-python-client
functions-framework
requests
google-cloud-firestore
langfuse
